[klock.vb]
Don't bother updating form, if running in system tray.

Added update of title and notifications for countdown, timer and
reminders.  This can be set for each.

Checks for data directory and creates if necessary.

[options.vb]

Added user options to add countdown, timer and reminder to form title and
notifications.

Added options to run klock on windows start up [and registry entries].

Added option to start klock minimised.

[selectTime.vb]

Amended fuzzytime so that one in the afternoon now read about oneish in
the afternoon [same for all hours].