[klock.vb]
Switched some of the error message box to display as notification windows.

[klock.iss]

Added un-install entry, so that registry items for the run on windows
start are removed.

[options.vb]

Fixed bug in test notification button code, called the notification
incorrectly.

[selectaction.vb]

Switched some of the error message box to display as notification windows.