[Klock.vb]
The data directory for storing the friends data, is not now hard coded -
it can be user configured.

[Options.vb]

Changes for the user selection of the friends file name and directory.

Added the ability to save friends file to Archive file [zip].

Added logic for setting the friends file name and directory.

Added logic for setting the friends achieve file name and directory.