Most of code for Events is now written - needs testing [lots].
[frmKlock.vb]

It seems that the timer maximum interval is a couple of minutes.

Now uses static variables to count the minutes.

More code for Events.

[frmNotification.vb]

Amended original code to run differnet for notification called as either
reminder of events.

Reminders will be closed by the time, but events will be closed by the
user.

[frmOptions.vb] & [UserSettings.vb]

More options added for Events.

[selectAction.vb]

Added new action = DisplayEvent()