Added the sayings tab.  This enables the displaying [via notifications] random sayings. The sayings are held in text files, which reside in a sayings directory within the application directory.  A set of sample text files and the directory are included.
[frmKlock.vb]
Added a timer for sayings.
Changed a number of if .. then .. else to if( , , )

[UserSettings.vb]
Added settings for sayings

[frmOptions.vb]
Logic to handle sayings settings.

[selectTime.vb]
Changed a number of if .. then .. else to if( , , )

[KlockThings.vb]
Added subs loadSayings() & randomSayings()
Changed return value of klockNotVisable()

[frmNotification.vb]
Amended Notification, so it is handled sayings.
Amended the mode - to be either R, S or E - for Reminder, Sayings or Event.
Removed the time-out from the input parameters.  Each type of notification now correctly sets the time-out according to the user settings for that type - within this module.
This meant a change to all instance where notifications are called throughout klock.

Added Hack-Regular.ttf to fonts directory.  A mono-spaced font used for list-boxes.

Ran spell checker across klock.