Added buttons to save and load the clipboard history data.
[IOcommon.vb]
Added logic to save and load the clipboard history in both CSB and BIN formats.

[frmClipboardMonitor.vb]
mainly code for above.
The listview was changed to a list box, so that it can hold objects like friends, events and memo.
Added code to change the fore colour of items in the list box - in listBoxThings.vb

[UserSettings.vb & frmOptions.vb]
Added settings for save clipboard as either CSV or bin.

[conversionThings.vb]
Improved the code in unitsLoad() - changed the way csv file are parsed.

[KlockThings.vb]
Removed itemForecolor() and moved back to frmClipboardMonitor.vb

[listBoxThings.vb]
Added ListBoxEnsureVisable(), which will scroll a list box so that the last entry is always visible.
Called on all list box’s - Friends, Events, Memo & Clipboard.

[clipItem.vb]
Added - class template for a clipboard item.