[frmBinaryKlock.vb] Added the ability to display the time as a Binary Klock. The Binary Klock has two modes, BCD and True Binary - selectable in user options.
[UserSettings.vb]
Added options for Binary Klock.

[frmBigTextKlock.vb]
Changed the fore colour of the status bar to that of the fore colour of the Big Text Klock.
Added logic for right click context menu.

[frmOptions.vb]
Added stuff for Binary Klock.

[frmSmallTextKlock.vb]
Changed the fore colour of the status bar to that of the fore colour of the Small Text Klock.
Added logic for right click context menu.

[KlockThings.vb]
Added Binary Klock Tool Tip to setToolTip()
Added Binary Klock to hot keys and re-ordered.
Added killChildKlocks()

[frmKlock.vb]
Added menu option to call Binary Klock.