[OptionThings.vb] Added file to hold some the boring stuff from frmOptions.vb.
[stickyNotes.vb]
Made the code more robust.

[frmOptions.vb]
Lost some stuff to OptionThings.vb.  Amended btnArchiveLoad() so that settings and sticky notes are loaded after files are loaded from an archive.  Removed the exclusion .log files from the archive.

[frmStickyNote.vb]
Added the label to the drag and drop routines.

[UserSettings.vb]
Removed usrTimeThreeQuartersChimes from options - not used.

[frmAbout.vb]
Added date and time of compile to version string.